service-timetable-python/requirements.txt
Flask
pandas
mysql-connector-python
openpyxl
python-dotenv